# backend/app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AlphaGBM API
  description: |
    股票智能分析平台 API 文档

    ## 认证方式
    所有需要认证的接口都需要在 HTTP Header 中携带 Supabase JWT Token:
    ```
    Authorization: Bearer <your_access_token>
    ```

    ## 服务类型
    | 服务类型 | 描述 |
    |---------|------|
    | stock_analysis | 股票分析 |
    | option_analysis | 期权分析 |
    | deep_report | 深度研报 |

    ## 投资风格
    | 风格 | 描述 |
    |------|------|
    | quality | 质量型 - 关注财务稳健、盈利能力强的优质公司 |
    | value | 价值型 - 寻找被市场低估的股票 |
    | growth | 成长型 - 追求高营收增长和盈利增长 |
    | momentum | 趋势型 - 跟随市场趋势和价格动量 |

  version: 1.0.0
  contact:
    name: AlphaGBM Team
    email: support@alphagbm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5002/api
    description: Development Server
  - url: https://api.alphagbm.com/api
    description: Production Server
...
paths:
...
  /payment/credits:
    get:
      tags: [Payment]
      summary: 获取用户额度
      description: 获取当前用户的可用额度信息
      operationId: getCredits
      security:
        - BearerAuth: []
      parameters:
        - name: service_type
          in: query
          required: false
          description: 服务类型
          schema:
            type: string
            enum: [stock_analysis, option_analysis, deep_report]
            default: stock_analysis
      responses:
        '200':
          description: 用户额度信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /payment/transactions:
    get:
      tags: [Payment]
      summary: 获取交易历史
      description: 获取当前用户的支付交易历史
      operationId: getTransactions
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          required: false
          description: 每页数量
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 交易历史
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /payment/create-checkout-session:
    post:
      tags: [Payment]
      summary: 创建支付会话
      description: 创建Stripe支付会话，用于购买订阅或加油包
      operationId: createCheckoutSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutRequest'
      responses:
        '200':
          description: 支付会话创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /payment/webhook:
    post:
      tags: [Payment]
      summary: Stripe Webhook回调
      description: |
        接收Stripe的Webhook回调，处理支付完成、订阅续费等事件。
        **注意：** 此接口由Stripe调用，无需手动调用
      operationId: handleWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook处理成功
        '400':
          description: 无效的Webhook请求
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT Token
  responses:
    Unauthorized:
      description: 未授权 - Token无效或已过期
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
  schemas:
...
    CreditsResponse:
      type: object
      properties:
        total_credits:
          type: integer
          description: 总剩余额度
        subscription:
          type: object
          nullable: true
          description: 订阅信息
        daily_free:
          type: object
          properties:
            quota:
              type: integer
              description: 每日免费配额
            used:
              type: integer
              description: 今日已使用
            remaining:
              type: integer
              description: 今日剩余
    TransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
              description:
                type: string
              amount:
                type: number
                format: float
              currency:
                type: string
              status:
                type: string
        total:
          type: integer
        pages:
          type: integer
        current_page:
          type: integer
    CreateCheckoutRequest:
      type: object
      required:
        - price_key
      properties:
        price_key:
          type: string
          description: 价格键
          enum: [plus_monthly, plus_yearly, pro_monthly, pro_yearly, topup_100]
        success_url:
          type: string
          description: 支付成功后跳转URL
        cancel_url:
          type: string
          description: 支付取消后跳转URL
    CheckoutSessionResponse:
      type: object
      properties:
        session_id:
          type: string
          description: Stripe会话ID
        checkout_url:
          type: string
          description: 支付页面URL